remove cyclic dep between MongoDb and queryfile
